+++ CHANGES.rst
@@ -6,8 +6,9 @@ Here you can see the full list of changes between pomp releases.
 Version 0.2.1
 -------------
 
-Not yet released
+Released at 2016-09-12
 
+- fix pipe start/stop/process exception processing
 - fix: `AioPomp` ensure_future for process_requests in main loop
 - `BaseCrawler.on_processing_done` when request, middlewares, response,
   extract, pipelines processing was done